Add files via upload
Poyan wants to make road He decided to choose whether the roads should be horizontal or vertical so that at the end the land is divided into as many parts as possible. For example, if n equals 3 and he draws three horizontal roads, the land is divided into 4 parts. But if he draws one horizontal road and two vertical roads, the land will be divided into 6 parts.

By taking the number n as input, say what is the maximum number of possible segments with n roads.